+++ src/migrations/m181009_120000_update_blacklist_settings.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace putyourlightson\snaptcha\migrations;
+
+use Craft;
+use craft\db\Migration;
+use putyourlightson\snaptcha\models\SettingsModel;
+use putyourlightson\snaptcha\Snaptcha;
+
+class m181009_120000_update_blacklist_settings extends Migration
+{
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    public function safeUp()
+    {
+        // Resave plugin settings converting blacklist to array of arrays for editable table field
+        /** @var SettingsModel $settings */
+        $settings = Snaptcha::$plugin->getSettings();
+
+        $blacklistedIps = preg_split('/\r\n|\r|\n/', $settings->blacklist);
+        $settings->blacklist = [];
+
+        foreach ($blacklistedIps as $blacklistedIp) {
+            $settings->blacklist[] = [$blacklistedIp];
+        }
+
+        Craft::$app->plugins->savePluginSettings(Snaptcha::$plugin, $settings->getAttributes());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown(): bool
+    {
+        echo "m181009_120000_update_blacklist_settings cannot be reverted.\n";
+
+        return false;
+    }
+}