+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # Changelog
 
+## 2.1.0 - Unreleased
+### Added
+- Added excluded URI patterns to settings
+- Added table field for blacklisted IP addresses in settings
+- Added check for `enableSnaptchaValidation` parameter in action's controller class
+
 ## 2.0.3 - 2018-10-08
+### Fixed
 - Fixed bug that was causing issues with live preview
 
 ## 2.0.2 - 2018-10-07
+### Fixed
 - Fixed bug that was preventing the correct error message from being displayed
 
 ## 2.0.1 - 2018-10-05
+### Fixed
 - Fixed bug with variable using incorrect StringHelper class
 
 ## 2.0.0 - 2018-10-04